refactor: packaging 수정
