+++ src/main/java/com/hongikbros/jobmanager/notice/query/applicaion/NoticeViewListService.java
@@ -1,17 +1,17 @@
 package com.hongikbros.jobmanager.notice.query.applicaion;
 
+import com.hongikbros.jobmanager.notice.query.applicaion.dto.NoticeResponses;
 import com.hongikbros.jobmanager.notice.query.dao.NoticeViewDao;
-import com.hongikbros.jobmanager.notice.query.dto.NoticeResponses;
 import com.hongikbros.jobmanager.security.core.CurrentMember;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-public class NoticeViewService {
+public class NoticeViewListService {
 
     private final NoticeViewDao noticeViewDao;
 
-    public NoticeViewService(NoticeViewDao noticeViewDao) {
+    public NoticeViewListService(NoticeViewDao noticeViewDao) {
         this.noticeViewDao = noticeViewDao;
     }
