+++ src/main/java/com/hongikbros/jobmanager/notice/query/dao/NoticeViewDao.java
@@ -1,6 +1,6 @@
 package com.hongikbros.jobmanager.notice.query.dao;
 
-import com.hongikbros.jobmanager.notice.command.dto.NoticeDetail;
+import com.hongikbros.jobmanager.notice.query.applicaion.dto.NoticeDetail;
 
 import java.util.List;
