+++ src/test/java/com/hongikbros/jobmanager/notice/documentation/NoticeDocumentationTest.java
@@ -6,12 +6,12 @@ import com.hongikbros.jobmanager.common.fixture.member.MemberFixture;
 import com.hongikbros.jobmanager.common.utils.TestObjectUtils;
 import com.hongikbros.jobmanager.member.domain.LoginMember;
 import com.hongikbros.jobmanager.notice.command.application.NoticeService;
+import com.hongikbros.jobmanager.notice.command.application.dto.NoticeCreateRequest;
 import com.hongikbros.jobmanager.notice.command.domain.notice.ApplyUrl;
 import com.hongikbros.jobmanager.notice.command.domain.notice.Company;
 import com.hongikbros.jobmanager.notice.command.domain.notice.Duration;
 import com.hongikbros.jobmanager.notice.command.domain.notice.Notice;
-import com.hongikbros.jobmanager.notice.command.dto.NoticeCreateRequest;
-import com.hongikbros.jobmanager.notice.command.dto.NoticeDetail;
+import com.hongikbros.jobmanager.notice.query.applicaion.dto.NoticeDetail;
 import com.hongikbros.jobmanager.notice.ui.NoticeController;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
